Move arithmetic functions to TimestampOp trait
