+++ arrow-array/src/types.rs
@@ -17,12 +17,7 @@
 
 //! Zero-sized types used to parameterize generic array implementations
 
-use crate::delta::{
-    add_days_datetime, add_months_datetime, shift_months, sub_days_datetime,
-    sub_months_datetime,
-};
-use crate::temporal_conversions::as_datetime_with_timezone;
-use crate::timezone::Tz;
+use crate::delta::shift_months;
 use crate::{ArrowNativeTypeOp, OffsetSizeTrait};
 use arrow_buffer::{i256, Buffer, OffsetBuffer};
 use arrow_data::decimal::{validate_decimal256_precision, validate_decimal_precision};
@@ -352,463 +347,6 @@ impl ArrowTimestampType for TimestampNanosecondType {
     fn make_value(naive: NaiveDateTime) -> Option<i64> {
         let nanos = naive.timestamp().checked_mul(1_000_000_000)?;
         nanos.checked_add(naive.timestamp_subsec_nanos() as i64)
-    }
-}
-
-fn add_year_months<T: ArrowTimestampType>(
-    timestamp: <T as ArrowPrimitiveType>::Native,
-    delta: <IntervalYearMonthType as ArrowPrimitiveType>::Native,
-    tz: Tz,
-) -> Option<<T as ArrowPrimitiveType>::Native> {
-    let months = IntervalYearMonthType::to_months(delta);
-    let res = as_datetime_with_timezone::<T>(timestamp, tz)?;
-    let res = add_months_datetime(res, months)?;
-    let res = res.naive_utc();
-    T::make_value(res)
-}
-
-fn add_day_time<T: ArrowTimestampType>(
-    timestamp: <T as ArrowPrimitiveType>::Native,
-    delta: <IntervalDayTimeType as ArrowPrimitiveType>::Native,
-    tz: Tz,
-) -> Option<<T as ArrowPrimitiveType>::Native> {
-    let (days, ms) = IntervalDayTimeType::to_parts(delta);
-    let res = as_datetime_with_timezone::<T>(timestamp, tz)?;
-    let res = add_days_datetime(res, days)?;
-    let res = res.checked_add_signed(Duration::milliseconds(ms as i64))?;
-    let res = res.naive_utc();
-    T::make_value(res)
-}
-
-fn add_month_day_nano<T: ArrowTimestampType>(
-    timestamp: <T as ArrowPrimitiveType>::Native,
-    delta: <IntervalMonthDayNanoType as ArrowPrimitiveType>::Native,
-    tz: Tz,
-) -> Option<<T as ArrowPrimitiveType>::Native> {
-    let (months, days, nanos) = IntervalMonthDayNanoType::to_parts(delta);
-    let res = as_datetime_with_timezone::<T>(timestamp, tz)?;
-    let res = add_months_datetime(res, months)?;
-    let res = add_days_datetime(res, days)?;
-    let res = res.checked_add_signed(Duration::nanoseconds(nanos))?;
-    let res = res.naive_utc();
-    T::make_value(res)
-}
-
-fn subtract_year_months<T: ArrowTimestampType>(
-    timestamp: <T as ArrowPrimitiveType>::Native,
-    delta: <IntervalYearMonthType as ArrowPrimitiveType>::Native,
-    tz: Tz,
-) -> Option<<T as ArrowPrimitiveType>::Native> {
-    let months = IntervalYearMonthType::to_months(delta);
-    let res = as_datetime_with_timezone::<T>(timestamp, tz)?;
-    let res = sub_months_datetime(res, months)?;
-    let res = res.naive_utc();
-    T::make_value(res)
-}
-
-fn subtract_day_time<T: ArrowTimestampType>(
-    timestamp: <T as ArrowPrimitiveType>::Native,
-    delta: <IntervalDayTimeType as ArrowPrimitiveType>::Native,
-    tz: Tz,
-) -> Option<<T as ArrowPrimitiveType>::Native> {
-    let (days, ms) = IntervalDayTimeType::to_parts(delta);
-    let res = as_datetime_with_timezone::<T>(timestamp, tz)?;
-    let res = sub_days_datetime(res, days)?;
-    let res = res.checked_sub_signed(Duration::milliseconds(ms as i64))?;
-    let res = res.naive_utc();
-    T::make_value(res)
-}
-
-fn subtract_month_day_nano<T: ArrowTimestampType>(
-    timestamp: <T as ArrowPrimitiveType>::Native,
-    delta: <IntervalMonthDayNanoType as ArrowPrimitiveType>::Native,
-    tz: Tz,
-) -> Option<<T as ArrowPrimitiveType>::Native> {
-    let (months, days, nanos) = IntervalMonthDayNanoType::to_parts(delta);
-    let res = as_datetime_with_timezone::<T>(timestamp, tz)?;
-    let res = sub_months_datetime(res, months)?;
-    let res = sub_days_datetime(res, days)?;
-    let res = res.checked_sub_signed(Duration::nanoseconds(nanos))?;
-    let res = res.naive_utc();
-    T::make_value(res)
-}
-
-impl TimestampSecondType {
-    /// Adds the given IntervalYearMonthType to an arrow TimestampSecondType.
-    ///
-    /// Returns `None` when it will result in overflow.
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn add_year_months(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalYearMonthType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        add_year_months::<Self>(timestamp, delta, tz)
-    }
-
-    /// Adds the given IntervalDayTimeType to an arrow TimestampSecondType.
-    ///
-    /// Returns `None` when it will result in overflow.
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn add_day_time(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalDayTimeType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        add_day_time::<Self>(timestamp, delta, tz)
-    }
-
-    /// Adds the given IntervalMonthDayNanoType to an arrow TimestampSecondType
-    ///
-    /// Returns `None` when it will result in overflow.
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn add_month_day_nano(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalMonthDayNanoType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        add_month_day_nano::<Self>(timestamp, delta, tz)
-    }
-
-    /// Subtracts the given IntervalYearMonthType to an arrow TimestampSecondType
-    ///
-    /// Returns `None` when it will result in overflow.
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn subtract_year_months(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalYearMonthType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        subtract_year_months::<Self>(timestamp, delta, tz)
-    }
-
-    /// Subtracts the given IntervalDayTimeType to an arrow TimestampSecondType
-    ///
-    /// Returns `None` when it will result in overflow.
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn subtract_day_time(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalDayTimeType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        subtract_day_time::<Self>(timestamp, delta, tz)
-    }
-
-    /// Subtracts the given IntervalMonthDayNanoType to an arrow TimestampSecondType
-    ///
-    /// Returns `None` when it will result in overflow.
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn subtract_month_day_nano(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalMonthDayNanoType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        subtract_month_day_nano::<Self>(timestamp, delta, tz)
-    }
-}
-
-impl TimestampMicrosecondType {
-    /// Adds the given IntervalYearMonthType to an arrow TimestampMicrosecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn add_year_months(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalYearMonthType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        add_year_months::<Self>(timestamp, delta, tz)
-    }
-
-    /// Adds the given IntervalDayTimeType to an arrow TimestampMicrosecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn add_day_time(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalDayTimeType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        add_day_time::<Self>(timestamp, delta, tz)
-    }
-
-    /// Adds the given IntervalMonthDayNanoType to an arrow TimestampMicrosecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn add_month_day_nano(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalMonthDayNanoType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        add_month_day_nano::<Self>(timestamp, delta, tz)
-    }
-
-    /// Subtracts the given IntervalYearMonthType to an arrow TimestampMicrosecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn subtract_year_months(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalYearMonthType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        subtract_year_months::<Self>(timestamp, delta, tz)
-    }
-
-    /// Subtracts the given IntervalDayTimeType to an arrow TimestampMicrosecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn subtract_day_time(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalDayTimeType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        subtract_day_time::<Self>(timestamp, delta, tz)
-    }
-
-    /// Subtracts the given IntervalMonthDayNanoType to an arrow TimestampMicrosecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn subtract_month_day_nano(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalMonthDayNanoType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        subtract_month_day_nano::<Self>(timestamp, delta, tz)
-    }
-}
-
-impl TimestampMillisecondType {
-    /// Adds the given IntervalYearMonthType to an arrow TimestampMillisecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn add_year_months(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalYearMonthType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        add_year_months::<Self>(timestamp, delta, tz)
-    }
-
-    /// Adds the given IntervalDayTimeType to an arrow TimestampMillisecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn add_day_time(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalDayTimeType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        add_day_time::<Self>(timestamp, delta, tz)
-    }
-
-    /// Adds the given IntervalMonthDayNanoType to an arrow TimestampMillisecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn add_month_day_nano(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalMonthDayNanoType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        add_month_day_nano::<Self>(timestamp, delta, tz)
-    }
-
-    /// Subtracts the given IntervalYearMonthType to an arrow TimestampMillisecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn subtract_year_months(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalYearMonthType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        subtract_year_months::<Self>(timestamp, delta, tz)
-    }
-
-    /// Subtracts the given IntervalDayTimeType to an arrow TimestampMillisecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn subtract_day_time(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalDayTimeType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        subtract_day_time::<Self>(timestamp, delta, tz)
-    }
-
-    /// Subtracts the given IntervalMonthDayNanoType to an arrow TimestampMillisecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn subtract_month_day_nano(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalMonthDayNanoType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        subtract_month_day_nano::<Self>(timestamp, delta, tz)
-    }
-}
-
-impl TimestampNanosecondType {
-    /// Adds the given IntervalYearMonthType to an arrow TimestampNanosecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn add_year_months(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalYearMonthType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        add_year_months::<Self>(timestamp, delta, tz)
-    }
-
-    /// Adds the given IntervalDayTimeType to an arrow TimestampNanosecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn add_day_time(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalDayTimeType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        add_day_time::<Self>(timestamp, delta, tz)
-    }
-
-    /// Adds the given IntervalMonthDayNanoType to an arrow TimestampNanosecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn add_month_day_nano(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalMonthDayNanoType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        add_month_day_nano::<Self>(timestamp, delta, tz)
-    }
-
-    /// Subtracts the given IntervalYearMonthType to an arrow TimestampNanosecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn subtract_year_months(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalYearMonthType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        subtract_year_months::<Self>(timestamp, delta, tz)
-    }
-
-    /// Subtracts the given IntervalDayTimeType to an arrow TimestampNanosecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn subtract_day_time(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalDayTimeType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        subtract_day_time::<Self>(timestamp, delta, tz)
-    }
-
-    /// Subtracts the given IntervalMonthDayNanoType to an arrow TimestampNanosecondType
-    ///
-    /// # Arguments
-    ///
-    /// * `timestamp` - The date on which to perform the operation
-    /// * `delta` - The interval to add
-    /// * `tz` - The timezone in which to interpret `timestamp`
-    pub fn subtract_month_day_nano(
-        timestamp: <Self as ArrowPrimitiveType>::Native,
-        delta: <IntervalMonthDayNanoType as ArrowPrimitiveType>::Native,
-        tz: Tz,
-    ) -> Option<<Self as ArrowPrimitiveType>::Native> {
-        subtract_month_day_nano::<Self>(timestamp, delta, tz)
     }
 }
